requirements.txt
flask==3.1.0
google-genai==1.10.0
Werkzeug==3.1.3
